+++ examples/stock-price-tool/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "pnpx bun src/mastra/index.ts"
   },
   "keywords": [],
   "author": "",
