+++ examples/stock-price-tool/src/mastra/agents/index.ts
