+++ examples/stock-price-tool/src/mastra/tools/index.ts
